+++ web/components/tipper.tsx
@@ -10,7 +10,7 @@ import { User } from 'common/user'
 import { formatMoney } from 'common/util/format'
 import { debounce, sumBy } from 'lodash'
 import { useEffect, useMemo, useRef, useState } from 'react'
-import { CommentTips } from 'web/hooks/use-comment-tips'
+import { CommentTips } from 'web/hooks/use-tip-txns'
 import { useUser } from 'web/hooks/use-user'
 import { transact } from 'web/lib/firebase/fn-call'
 import { Row } from './layout/row'
